requirements-dev.txt
attrs==21.4.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cloudevents==1.2.0
deprecation==2.1.0
Flask==2.1.2
functions-framework==3.0.0
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-bigquery==3.1.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
gunicorn==20.1.0
httpretty==1.1.4
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mock==4.0.3
numpy==1.22.4
packaging==21.3
pluggy==1.0.0
proto-plus==1.20.5
protobuf==3.20.1
py==1.11.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
watchdog==1.0.2
Werkzeug==2.1.2
zipp==3.8.0
